Pridani editace profilu, rozdeli user controlleru na user a authentication controller
