+++ API/config/dbconnection.js
@@ -3,9 +3,10 @@ const mysql = require('mysql');
 var connection = mysql.createConnection({
 	host: 'localhost',
 	user: 'root',
-	password: '',
+	password: 'root',
 	database: 'iis-db',
-	dateStrings: true
+	dateStrings: true,
+	socketPath: '/Applications/MAMP/tmp/mysql/mysql.sock'
 });
 
 connection.connect(function(err) {
